Refactor wayland input processing
SeatInputProcessor has become too big
