+++ xbmc/windowing/wayland/CMakeLists.txt
@@ -6,6 +6,9 @@ set_source_files_properties(${WAYLAND_EXTRA_PROTOCOL_GENERATED_DIR}/wayland-extr
 set(SOURCES Connection.cpp
             Output.cpp
             GLContextEGL.cpp
+            InputProcessorKeyboard.h
+            InputProcessorPointer.h
+            InputProcessorTouch.h
             OSScreenSaverIdleInhibitUnstableV1.cpp
             SeatInputProcessor.cpp
             ShellSurface.cpp
@@ -19,6 +22,9 @@ set(SOURCES Connection.cpp
 set(HEADERS Connection.h
             Output.h
             GLContextEGL.h
+            InputProcessorKeyboard.cpp
+            InputProcessorPointer.cpp
+            InputProcessorTouch.cpp
             OSScreenSaverIdleInhibitUnstableV1.h
             SeatInputProcessor.h
             ShellSurface.h
